client-settings refactor renombre
