+++ app/resources/client-settings/client-settings.js
